# swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/health":{"get":{"operationId":"HealthController_get","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/metrics":{"get":{"operationId":"MetricsController_metrics","parameters":[],"responses":{"200":{"description":""}}}},"/api/v1/internal/cache/metrics":{"get":{"operationId":"InternalController_metrics","summary":"缓存命中率与配额指标","parameters":[],"responses":{"200":{"description":"返回 DSL 缓存服务的命中率与配额统计"}},"tags":["internal"],"security":[{"bearer":[]}]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/api/v1/users/me":{"get":{"operationId":"UsersController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/api/v1/projects/{projectId}/tables":{"post":{"operationId":"TablesController_create","summary":"创建表","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTableDto"}}}},"responses":{"200":{"description":"创建成功返回表实体"}},"tags":["tables"],"security":[{"bearer":[]}]},"get":{"operationId":"TablesController_list","summary":"列出表（只读，支持匿名）","parameters":[{"name":"projectId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"表列表"}},"tags":["tables"],"security":[{"bearer":[]}]}},"/api/v1/projects/{projectId}/tables/{tableId}":{"patch":{"operationId":"TablesController_update","summary":"更新表属性（名称/匿名只读开关）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTableDto"}}}},"responses":{"200":{"description":"更新成功返回表实体"}},"tags":["tables"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/fields":{"post":{"operationId":"FieldsController_create","summary":"新增字段","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateFieldDto"}}}},"responses":{"200":{"description":"创建成功返回字段实体"}},"tags":["fields"],"security":[{"bearer":[]}]},"get":{"operationId":"FieldsController_list","summary":"列出字段（只读，支持匿名；匿名按表开关与字段权限过滤）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"字段列表"}},"tags":["fields"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/fields/{fieldId}":{"patch":{"operationId":"FieldsController_update","summary":"更新字段（名称/类型/描述/可见性/顺序/权限）","parameters":[{"name":"fieldId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateFieldDto"}}}},"responses":{"200":{"description":"更新成功返回字段实体"}},"tags":["fields"],"security":[{"bearer":[]}]},"delete":{"operationId":"FieldsController_remove","summary":"删除字段","parameters":[{"name":"fieldId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"删除成功返回删除的字段实体"}},"tags":["fields"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/views":{"get":{"operationId":"ViewsController_listByTable","summary":"列出表的视图（只读，支持匿名；匿名受表开关限制）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"视图列表"}},"tags":["views"],"security":[{"bearer":[]}]},"post":{"operationId":"ViewsController_createView","summary":"创建视图（需登录，编辑者及以上）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateViewDto"},"examples":{"default":{"summary":"创建视图示例","value":{"name":"我的视图","config":{"columnVisibility":{"field-text":true,"field-number":true},"sorting":[{"fieldId":"field-text","dir":"asc"},{"fieldId":"field-number","dir":"desc"}],"groupBy":"field-date","groupDateGranularity":"month","aggregations":[{"type":"count"},{"type":"sum","fieldId":"field-number"}],"viewPriority":false,"freezeCount":1,"rowHeight":28,"kanbanGroupFieldId":"field-status","calendarFields":{"startDateFieldId":"start","endDateFieldId":"end"},"sharedRoles":["VIEWER","EDITOR"],"exportEnabled":false}}}}}}},"responses":{"200":{"description":"创建成功返回视图"}},"tags":["views"],"security":[{"bearer":[]}]}},"/api/v1/views/{viewId}":{"get":{"operationId":"ViewsController_getOne","summary":"获取指定视图（套用访问守卫）","parameters":[{"name":"viewId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"视图详情"}},"tags":["views"],"security":[{"bearer":[]}]},"patch":{"operationId":"ViewsController_updateView","summary":"更新视图（需登录，编辑者及以上）","parameters":[{"name":"viewId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateViewDto"},"examples":{"enableExportForEditors":{"summary":"开启导出并限制共享角色","value":{"name":"报表视图","config":{"sharedRoles":["EDITOR","ADMIN"],"exportEnabled":true,"sorting":[{"fieldId":"field-date","dir":"desc"}],"aggregations":[{"type":"avg","fieldId":"field-number"}]}}}}}}},"responses":{"200":{"description":"更新成功返回视图"}},"tags":["views"],"security":[{"bearer":[]}]}},"/api/v1/projects":{"post":{"operationId":"ProjectsController_create","summary":"创建项目（支持设置匿名只读开关）","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectDto"}}}},"responses":{"200":{"description":"创建成功返回项目实体"}},"tags":["projects"],"security":[{"bearer":[]}]}},"/api/v1/projects/{projectId}":{"patch":{"operationId":"ProjectsController_update","summary":"更新项目（名称/匿名只读开关）","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectDto"}}}},"responses":{"200":{"description":"更新成功返回项目实体"}},"tags":["projects"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/records":{"post":{"operationId":"RecordsController_create","summary":"新增记录","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRecordDto"}}}},"responses":{"200":{"description":"创建成功返回记录及数据"}},"tags":["records"],"security":[{"bearer":[]}]},"get":{"operationId":"RecordsController_list","summary":"查询记录（分页、基本过滤，支持匿名）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"example":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"example":20,"type":"number"}},{"name":"q","required":false,"in":"query","description":"简单文本过滤，匹配 valueText","schema":{"example":"keyword","type":"string"}}],"responses":{"200":{"description":"记录列表"}},"tags":["records"],"security":[{"bearer":[]}]}},"/api/v1/records/{recordId}":{"patch":{"operationId":"RecordsController_update","summary":"局部更新记录","parameters":[{"name":"recordId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRecordDto"}}}},"responses":{"200":{"description":"更新成功返回记录ID"}},"tags":["records"],"security":[{"bearer":[]}]},"delete":{"operationId":"RecordsController_remove","summary":"删除记录","parameters":[{"name":"recordId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"删除成功返回记录ID"}},"tags":["records"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/records{batch}":{"post":{"operationId":"RecordsController_batch","summary":"批量新增/更新/删除（事务）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchRecordsDto"}}}},"responses":{"200":{"description":"批量操作结果"}},"tags":["records"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/query":{"post":{"operationId":"QueryController_query","summary":"DSL 查询（过滤/排序/分组/聚合，支持匿名）","parameters":[{"name":"tableId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequestDto"}}}},"responses":{"200":{"description":"查询结果"}},"tags":["query"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/export":{"get":{"operationId":"ExportController_export","summary":"按视图导出记录为 Excel/CSV 文件","parameters":[{"name":"tableId","required":true,"in":"path","description":"表ID","schema":{"type":"string"}},{"name":"viewId","required":true,"in":"query","description":"视图ID","schema":{"type":"string"}},{"name":"format","required":false,"in":"query","description":"文件格式：xlsx/csv（默认xlsx）","schema":{"type":"string"}}],"responses":{"200":{"description":"Excel/CSV 文件下载"}},"tags":["export"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/import":{"post":{"operationId":"ImportController_import","summary":"上传并导入为新表（CSV/XLSX）","parameters":[{"name":"tableId","required":true,"in":"path","description":"参考表ID（用其项目ID）","schema":{"type":"string"}}],"responses":{"200":{"description":"返回新建表ID与导入统计"}},"tags":["import"],"security":[{"bearer":[]}]}},"/api/v1/tables/{tableId}/attachments":{"post":{"operationId":"AttachmentsController_upload","summary":"上传附件（校验配额并持久化）","parameters":[{"name":"tableId","required":true,"in":"path","description":"表ID","schema":{"type":"string"}}],"responses":{"200":{"description":"上传成功返回附件记录"}},"tags":["attachments"],"security":[{"bearer":[]}]}},"/api/v1/home/recent-tables":{"get":{"operationId":"HomeController_recentTables","summary":"最近编辑过的数据表列表（按记录更新排序）","parameters":[],"responses":{"200":{"description":"返回最近有更新记录的表列表（去重）"}},"tags":["home"],"security":[{"bearer":[]}]}},"/api/v1/space/my":{"get":{"operationId":"SpaceController_mySpace","summary":"我的空间：项目/任务/表层级（任务占位）","parameters":[],"responses":{"200":{"description":"返回当前用户拥有的项目及其下数据表（任务为空占位）"}},"tags":["space"],"security":[{"bearer":[]}]}},"/api/v1/space/project":{"get":{"operationId":"SpaceController_projectSpace","summary":"项目空间：公共视图表（匿名只读开关）","parameters":[],"responses":{"200":{"description":"返回允许匿名只读的项目及其允许匿名的表"}},"tags":["space"],"security":[{"bearer":[]}]}}},"info":{"title":"Sheets Backend API","description":"API documentation for sheets backend","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{}},"CreateTableDto":{"type":"object","properties":{"name":{"type":"string","example":"任务表"},"isAnonymousReadEnabled":{"type":"boolean","example":false,"description":"匿名只读开关"}},"required":["name"]},"UpdateTableDto":{"type":"object","properties":{"name":{"type":"string","example":"新的任务表名"},"isAnonymousReadEnabled":{"type":"boolean","example":true,"description":"匿名只读开关"}}},"CreateFieldDto":{"type":"object","properties":{"name":{"type":"string","example":"标题"},"type":{"type":"string","enum":["TEXT","NUMBER","DATE"],"example":"TEXT"},"description":{"type":"string","example":"任务标题"},"visible":{"type":"boolean","example":true},"order":{"type":"number","example":1},"permissionJson":{"type":"object","example":{"visibility":"visible","writableRoles":["EDITOR","ADMIN","OWNER"]}}},"required":["name","type"]},"UpdateFieldDto":{"type":"object","properties":{"name":{"type":"string","example":"新的标题"},"type":{"type":"string","enum":["TEXT","NUMBER","DATE"],"example":"TEXT"},"description":{"type":"string","example":"描述更新"},"visible":{"type":"boolean","example":true},"order":{"type":"number","example":3},"permissionJson":{"type":"object","example":{"visibility":"readonly","readableRoles":["VIEWER","EDITOR","ADMIN","OWNER"]}}}},"ViewSortingItemDto":{"type":"object","properties":{"id":{"type":"string","description":"字段ID"},"desc":{"type":"boolean","description":"是否降序","default":false}}},"ViewAggregationDto":{"type":"object","properties":{"type":{"type":"string","enum":["count","sum","avg","min","max"]},"fieldId":{"type":"string","description":"聚合字段ID（count可省略）"}}},"ViewConfigDto":{"type":"object","properties":{"columnVisibility":{"type":"object","description":"列可见性：key为字段ID，false表示隐藏"},"sorting":{"description":"排序配置","type":"array","items":{"$ref":"#/components/schemas/ViewSortingItemDto"}},"groupBy":{"type":"string","description":"分组字段ID"},"groupDateGranularity":{"type":"string","description":"日期分组粒度","enum":["day","week","month"]},"aggregations":{"description":"聚合定义","type":"array","items":{"$ref":"#/components/schemas/ViewAggregationDto"}},"viewPriority":{"type":"boolean","description":"视图优先：服务端按视图覆盖客户端查询参数","default":false},"freezeCount":{"type":"number","description":"冻结列数量","default":0},"rowHeight":{"type":"number","description":"行高（像素）","default":28},"kanbanGroupFieldId":{"type":"string","description":"看板分组字段ID"},"calendarFields":{"type":"object","description":"日历字段设置，如开始/结束日期"},"sharedRoles":{"type":"array","description":"共享角色列表（允许访问该视图的角色）","items":{"type":"string","enum":["VIEWER","EDITOR","ADMIN","OWNER"]}},"exportEnabled":{"type":"boolean","description":"允许导出（默认 false，编辑者及以上可开启）","default":false}}},"CreateViewDto":{"type":"object","properties":{"name":{"type":"string","description":"视图名称","example":"我的视图"},"config":{"description":"视图配置","allOf":[{"$ref":"#/components/schemas/ViewConfigDto"}]}}},"UpdateViewDto":{"type":"object","properties":{"name":{"type":"string","description":"视图名称","example":"重命名视图"},"config":{"description":"视图配置（列可见性/排序/分组/聚合等）","allOf":[{"$ref":"#/components/schemas/ViewConfigDto"}]}}},"CreateProjectDto":{"type":"object","properties":{"name":{"type":"string","example":"我的项目"},"isAnonymousReadEnabled":{"type":"boolean","example":false,"description":"项目级匿名只读开关"}},"required":["name"]},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string","example":"新的项目名称"},"isAnonymousReadEnabled":{"type":"boolean","example":true,"description":"项目级匿名只读开关"}}},"CreateRecordDto":{"type":"object","properties":{"data":{"type":"object","description":"记录数据，键为字段ID","example":{"field-uuid-1":"文本","field-uuid-2":123}}}},"UpdateRecordDto":{"type":"object","properties":{"data":{"type":"object","description":"局部更新数据，键为字段ID","example":{"field-uuid-1":"新文本"}}},"required":["data"]},"BatchCreateItemDto":{"type":"object","properties":{}},"BatchUpdateItemDto":{"type":"object","properties":{}},"BatchRecordsDto":{"type":"object","properties":{"create":{"type":"array","items":{"$ref":"#/components/schemas/BatchCreateItemDto"}},"update":{"type":"array","items":{"$ref":"#/components/schemas/BatchUpdateItemDto"}},"delete":{"type":"array","items":{"type":"string"}}}},"QueryFilterDto":{"type":"object","properties":{"fieldId":{"type":"string","description":"字段ID"},"op":{"type":"string","enum":["eq","neq","lt","gt","lte","gte","in","like","regex","between","exists"]},"value":{"type":"object","description":"单值或区间起始值"},"values":{"description":"多值列表或区间结束值","type":"array","items":{"type":"string"}}}},"QuerySortDto":{"type":"object","properties":{"fieldId":{"type":"string","description":"字段ID"},"direction":{"type":"string","enum":["asc","desc"]}}},"QueryRequestDto":{"type":"object","properties":{"filters":{"type":"array","items":{"$ref":"#/components/schemas/QueryFilterDto"}},"filterLogic":{"type":"string","description":"过滤组合逻辑","enum":["and","or"]},"sorts":{"type":"array","items":{"$ref":"#/components/schemas/QuerySortDto"}},"groupBy":{"type":"string","description":"分组字段ID"},"groupDateGranularity":{"type":"string","description":"日期分组粒度","enum":["day","week","month"]},"aggregations":{"description":"聚合定义，如 count/sum/avg/min/max","example":[{"type":"count"},{"type":"sum","fieldId":"fid-number"}],"type":"array","items":{"type":"string"}},"viewId":{"type":"string","description":"关联视图ID（可选）"},"page":{"type":"number","description":"页码","example":1},"pageSize":{"type":"number","description":"每页数量","example":20},"useCache":{"type":"boolean","description":"是否启用缓存（认证用户可选）","example":false},"cacheTtlMs":{"type":"number","description":"缓存 TTL 毫秒（上限受服务端约束）","example":15000}}}}}}